requirements.txt
actdiag==3.0.0
alabaster==0.7.13
asyncgui==0.6.3
asynckivy==0.6.4
Babel==2.12.1
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
blockdiag==3.0.0
blurb==1.2.1
build==1.2.1
cachetools==5.5.0
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Cython==3.0.11
distlib==0.3.8
docutils==0.16
domdf-python-tools==3.9.0
exceptiongroup==1.2.2
filelock==3.15.4
funcparserlib==2.0.0a0
furo==2023.9.10
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.4
imagesize==1.4.1
iso8601==0.1.12
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jaraco.packaging==10.2.3
jaraco.path==3.7.1
jaraco.vcs==2.4.0
jaraco.versioning==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
materialyoucolor==2.0.9
more-itertools==10.4.0
natsort==8.4.0
nwdiag==3.0.0
oauthlib==3.2.2
packaging==23.1
pillow==10.4.0
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pychallonge==1.11.5
Pygments==2.15.1
pygments-github-lexers==0.0.5
pyparsing==3.1.4
pyproject_hooks==1.1.0
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
pyqt6-tools==6.4.2.3.3
PyQt6_sip==13.8.0
python-dateutil==2.9.0.post0
python-docs-theme==2024.6
python-dotenv==1.0.1
pytz==2019.3
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
rst.linker==2.6.0
seqdiag==3.0.0
settingsjson==1.0.2
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==3.2.1
sphinx-basic-ng==1.0.0b2
sphinxcontrib-actdiag==3.0.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-blockdiag==3.0.0
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-nwdiag==2.0.0
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-seqdiag==3.0.0
sphinxcontrib-serializinghtml==1.1.5
sphinxext-opengraph==0.9.1
tempora==5.7.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2020.3
tzlocal==2.1
uritemplate==4.1.1
urllib3==2.0.4
virtualenv==20.26.3
webcolors==24.8.0
